refactor: Refatorar o tratamento e a validação de depoimentos, atualizar esquemas e serviços.
