+++ app/repositories/testimonial_repository.py
@@ -1,9 +1,9 @@
 from app.database.supabase import get_supabase_client
-from app.schemas.testimonial import TestimonialsSchema
+from app.schemas.testimonial import TestimonialSchema
 
 supabase = get_supabase_client()
 
-def create_testimonial_repository(testimonial: TestimonialsSchema):
+def create_testimonial_repository(testimonial: TestimonialSchema):
     response = (supabase.table("testimonials").insert(testimonial.model_dump(mode="json")).execute())
 
     return response.data[0]
